// src/interfaces/StoreTheme.tsx
export interface StoreTheme {
    light: {
        pageBackground: string;
        pageText: string;
        buttonBackground: string;
        buttonText: string;
        buttonShadow: string;
        tint: string;
        tabIconDefault: string;
        tabIconSelected: string;
        cardBackground: string;
        cardText: string;
        cardSecondaryText: string;
        cardPrice: string;
        cardBorder: string;
        background: string;
        title: string;
        subtitle: string;
        border: string;
        card: string;
        primary: string;
        secondary: string;
    };
    dark: {
        pageBackground: string;
        pageText: string;
        buttonBackground: string;
        buttonText: string;
        buttonShadow: string;
        tint: string;
        tabIconDefault: string;
        tabIconSelected: string;
        cardBackground: string;
        cardText: string;
        cardSecondaryText: string;
        cardPrice: string;
        cardBorder: string;
        background: string;
        title: string;
        subtitle: string;
        border: string;
        card: string;
        primary: string;
        secondary: string;
    };
}